Fix NavBar & SystemBar sizing issues
PhoneWindowManager makes some decisions regarding
portrait & landscape based on screen sizes.  Since we poke
this routine everytime we change something on the NavBar, this
can result in PWN changing the orientation of the device.  This
was causing strange behavior in resizing of the NavBar & SystemBar.
It might also have been causing other strange orientation issues.

PS2:  Attempts to decode the actual device type based on the screen
   dimensions and the rotation at time of setInitialDisplaySize.
   This is needed because large tablets (that default to a landscape
   orientation, need to have the dimensions swapped)

PS3: Whitespace

ps 4: Restore the stock calculations for orientation, but restrict
    them to running only one time (at initial boot).  This will allow
    the normal detection routines to set up the proper rotation variables,
    but will not reset (and change them incorrectly) everytime we poke
   setInitialDisplaySize.

Signed-off-by: Zaphod <mwilson@acesappraisal.com>
Change-Id: Iaec0e8ba623f10bd0d4118b4d5a99a42ab0564f6

Conflicts:

	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java